Electricity Board Program: Write a program to calculate electricity charges based on consumption.
An electricity board charges the following rates to domestic users to discourage
large consumption of energy.
For the first 100 units: - 60 P per unit
For the next 200 units: -80 P per unit
 Beyond 300 units: -90 P per unit
All users are charged a minimum of Rs 50 if the total amount is more than Rs 300
then an additional surcharge of 15% is added.